Preliminar support for php72
With sqlsrv latest 5.2-RC1 drivers and MSODBC 17 preview
it seems that it's possible to make the extension to work
under Debian 9.

it requires to install MSODBC 13, uninstall it and then install
MSODBC 17, but seems to be working ok.

oci also requires some little changes, have been updated to
12.2 libraries and is using proper ENV for both interactive
and non-interactive sessions.

and, finally, what I've been unable to fix/make work is solr. it
seems that still is not ready, so I've disabled it for now.

This will help fixing #16.